Icons svg moved to icons folder
