+++ frontend/src/pages/Auth/ForgotPassword/index.tsx
@@ -3,7 +3,7 @@ import { AiOutlineMail } from "react-icons/ai";
 import { Link, useNavigate } from "react-router-dom";
 import { Button } from "../../../components/AuthFormComponents/Button";
 import { Input } from "../../../components/AuthFormComponents/Input";
-import LogoSvg from "../../../components/Logo";
+import LogoSvg from "../../../components/Icons/Logo";
 import { sendPasswordResetEmail } from "firebase/auth";
 import { auth } from "../../../config/firebase";
